+++ packages/amplify_core/lib/src/types/notifications/push/remote_message_callback.dart
@@ -4,5 +4,5 @@
 import '../notification_types.dart';
 
 typedef RemoteMessageCallback = void Function(
-  RemotePushMessage remotePushMessage,
+  PushNotificationMessage remotePushMessage,
 );