+++ packages/notifications/push/amplify_push_notifications_pinpoint/lib/pinpoint_provider.dart
@@ -3,5 +3,23 @@
 
 import 'package:amplify_core/amplify_core.dart';
 
-// TODO: Complete the implementation using Analytics Client
-class PinpointProvider extends ServiceProviderClient {}
+// TODO(Samaritan1011001): Complete the implementation using Analytics Client
+class PinpointProvider extends ServiceProviderClient {
+  @override
+  void init({
+    AmplifyConfig? config,
+    required AmplifyAuthProviderRepository authProviderRepo,
+  }) {
+    // TODO(Samaritan1011001): implement init
+  }
+
+  @override
+  void recordNotificationEvent({required AnalyticsEvent event}) {
+    // TODO(Samaritan1011001): implement recordNotificationEvent
+  }
+
+  @override
+  void registerDevice(String deviceToken) {
+    // TODO(Samaritan1011001): implement registerDevice
+  }
+}